+++ src/main/java/hello/configuration/OAuth2ServerConfiguration.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package hello;
+package hello.configuration;
 
+import hello.service.CustomUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
