+++ src/main/java/hello/configuration/WebSecurityConfiguration.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package hello;
+package hello.configuration;
 
+import hello.service.CustomUserDetailsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
