+++ src/main/java/hello/controllers/GreetingController.java
@@ -14,17 +14,27 @@
  * limitations under the License.
  */
 
-package hello;
+package hello.controllers;
 
+import hello.models.Greeting;
+import hello.data.User;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+import org.springframework.security.web.bind.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class HomeController {
+public class GreetingController {
 
-	@RequestMapping("/")
-	public String home() {
-		return "home";
+	private static final String template = "Hello, %s!";
+	
+	private final AtomicLong counter = new AtomicLong();
+
+	@RequestMapping("/greeting")
+	public Greeting greeting(@AuthenticationPrincipal User user) {
+		return new Greeting(counter.incrementAndGet(), String.format(template, user.getName()));
 	}
 
 }
