+++ src/test/java/hello/GreetingControllerTest.java
@@ -16,6 +16,7 @@
 
 package hello;
 
+import hello.controllers.GreetingController;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
